* develop: (522 commits)
  fixed some typos
  0.8.12
  set version number
  adding the sbt-pgp plugin locally, does not make sense…
  added “attribute” class
  updated changes
  fixed the handling of “Formal Parameters” (nameless parameters)
  complete overhaul how (type) annotations are represented
  we are now using “:” for separating values from their descriptors
  improved the formatting for lists
  setting the utf8 file encoding within the sbtops is not enough on all installations.
  added annotation to an instance of type check
  added the pgp plugin
  added a test case related to anntoations of types of intersection types used in casts
  added the sonatype plugin to OPAL (the gpg is still externally configured…)
  fixed toXHTML for Type_Argument_Target subtypes
  improved the test fixture related to type annotations
  added a link to the JSR docs
  fixed formatting violation
  clarrified the “one commit” and “no pull requests w.r.t. the master branch” rules
  ...

# Conflicts:
#	OPAL/ai/src/main/scala/org/opalj/tac/AsQuadruples.scala
#	OPAL/ai/src/main/scala/org/opalj/tac/Expr.scala
#	OPAL/ai/src/main/scala/org/opalj/tac/Stmt.scala
#	OPAL/ai/src/main/scala/org/opalj/tac/ToJavaLike.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/ArrayTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/CastTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/CmpToZeroIfTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/ConstantsTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/DoubleArithmeticTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/FloatArithmeticTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/IntegerArithmeticTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/IntegerIfTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/LongArithmeticTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/RefCmpIfTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/StackAndSynchronizationTest.scala
#	OPAL/ai/src/test/scala/org/opalj/tac/SwitchTest.scala
#	OPAL/br/src/main/scala/org/opalj/br/ClassHierarchy.scala

+++ OPAL/common/src/main/scala/org/opalj/graphs/MutableNodeLike.scala
@@ -1,5 +1,5 @@
 /* BSD 2-Clause License:
- * Copyright (c) 2009 - 2016
+ * Copyright (c) 2009 - 2017
  * Software Technology Group
  * Department of Computer Science
  * Technische Universität Darmstadt
@@ -38,18 +38,19 @@ import scala.collection.Map
  * ==Thread Safety==
  * This class is thread-safe. It is possible to add multiple child nodes concurrently.
  *
- * @see The demo project for example usages.
+ * @see    The demo project for example usages.
  *
  * @tparam I The type of the object that is associated with this node/the type of
- *      the object for which this node object is created.
- * @param identifier The underlying object.
- *       '''The underlying object must correctly implement the equals/hashCode contract.
- *       I.e., the `hashCode` of two object instances that are added to the same graph is
- *      different whenever `equals` is `false`'''.
- * @param identifierToString A function that converts "an" identifier to a string. By
- *      default the given object's `toString` method is called. It is possible
- *      that a graph has two nodes with the same textual representation representation
- *      but a different identity.
+ *         the object for which this node object is created.
+ * @param  theIdentifier The underlying object. '''For all nodes of a graph the hashCode method
+ *         has to return a unique id unless the nodeId method is overridden.'''
+ *         '''The underlying object must correctly implement the equals/hashCode contract.
+ *         I.e., the `hashCode` of two object instances that are added to the same graph is
+ *         different whenever `equals` is `false`'''.
+ * @param  identifierToString A function that converts "an" identifier to a string. By
+ *         default the given object's `toString` method is called. It is possible
+ *         that a graph has two nodes with the same textual representation representation
+ *         but a different identity.
  * @author Michael Eichberg
  */
 class MutableNodeLike[I, N <: Node](
@@ -73,7 +74,7 @@ class MutableNodeLike[I, N <: Node](
 
     override def toHRR: Option[String] = Some(identifierToString(identifier))
 
-    override def nodeId: Long = identifier.hashCode().toLong
+    override def nodeId: Int = identifier.hashCode()
 
     def addChild(node: N): Unit = this.synchronized(theChildren = node :: children)
